Changed static to dynamic arrays appropriately
There were a few places where we need to keep track of everything that
is allocated (and any number of things can be allocated) until ADB
closes.
Before everything was statically allocated with a large number, but this
could theoretically not be enough memory.
Added a polymorphic array list data structure instead (which acts
similarly to a Java ArrayList). Declaring the variables is a little
awkward, but it is easy to use them.